main.py
import yaml
from flask import Flask, request, jsonify
from google.cloud import asset_v1 
from google.cloud import bigquery
import vertexai
from vertexai.language_models import TextGenerationModel

# ... (Additional imports if needed)

app = Flask(__name__)

# Placeholder for establishing database connections
client = bigquery.Client()
asset_client = asset_v1.AssetServiceClient()
#vertex_ai_client = aiplatform_v1.Client()

               
def k8_deployment_parse(k8_object, project, cluster_name, kind):   
    containers = k8_object['spec']['template']['spec']['containers']
  
    for i in range(len(containers)):
        #print(k8_object['spec']['template']['spec']['containers'][i]['name'])
        #print(k8_object['spec']['template']['spec']['containers'][i]['image'])
        #print(k8_object['spec']['template']['spec']['containers'][i]['resources']['requests']['cpu'])
        #print(k8_object['spec']['template']['spec']['containers'][i]['resources']['limits']['cpu'])
        #print(k8_object['spec']['template']['spec']['containers'][i]['resources']['requests']['memory'])
        #print(k8_object['spec']['template']['spec']['containers'][i]['resources']['limits']['memory'])
        k8_details = {
            "project_id": project,
            "cluster_name": cluster_name,
            "controller_name": k8_object['metadata']['name'],
            "controller_type": kind,
            "namespace_name": k8_object['metadata']['namespace'],
            "container_name": k8_object['spec']['template']['spec']['containers'][i]['name']
        }
    return k8_details

def determine_qos(resource_requests, resource_limits):
    """Determines Quality of Service (QoS) based on resource requests/limits.

    Args:
        resource_requests: Resource requests from the YAML file.
        resource_limits: Resource limits from the YAML file.

    Returns:
        A string representing the QoS category ('guaranteed', 'burstable', or 'best_effort').
    """

    # ... Logic to implement QoS determination
    return qos

def check_workload_exists(project_id, cluster_name, container_info):
    """Query BigQuery to check if a record exists
    Returns:
        true if the workload exists in BigQuery.
    """
    from google.cloud import bigquery
    print(container_info)

    container_name = container_info["container_name"]
    namespace_name = container_info["namespace_name"]
    controller_name = container_info["controller_name"]
    controller_type = container_info["controller_type"]
    
    query = f"""
        SELECT 
        COUNT(name) as found
        FROM `gke-opt-demo.1034414536999_Metrics._AllMetrics` 
        WHERE name = 'kubernetes.io/container/cpu/core_usage_time' AND
        resource.labels.cluster_name = '{cluster_name}' AND
        resource.labels.container_name = "{container_name}" AND
        resource.labels.namespace_name = {namespace_name} AND
        system_labels.top_level_controller_name = {controller_name} AND
        system_labels.top_level_controller_type = {controller_type}
        LIMIT 1
    """
    print(query)
    quit()
    # Construct a BigQuery client object.
    client = bigquery.Client()
    results = client.query_and_wait(
    """
        SELECT 
        COUNT(name) as found
        FROM `gke-opt-demo.1034414536999_Metrics._AllMetrics` 
        WHERE name = 'kubernetes.io/container/cpu/core_usage_time' AND
        resource.labels.cluster_name = "{cluster_name}" AND
        resource.labels.container_name = '{container_name}' AND
        resource.labels.namespace_name = '{namespace_name}' AND
        system_labels.top_level_controller_name = '{controller_name}' AND
        system_labels.top_level_controller_type = '{controller_type}'
        LIMIT 1
    """)
    for row in results:
        print(row)
    
    return true

def get_workload_recommendation(project_id, cluster_name, container_info):
    """Executes a BigQuery query.

    Args:
        query: The SQL query string.

    Returns:
        The results of the BigQuery query.
    """

    # ... Logic to execute the query using the 'client'
    return query_results


def call_vertex_ai(image_name):
    """Calls the Vertex AI API to analyze an image.

    Returns:
        Information about the image type and instance recommendations.
    """

    # ... Logic to make the API call using 'vertex_ai_client'
    return image_info

def call_asset_inventory():
    """Extracts relevant information from the provided YAML file.

    Args:
       yaml_data: The content of the YAML file.

    Returns:
        A dictionary containing the extracted information.
    """
    return 0



@app.route('/api/optimize', methods=['POST'])
def optimize():
    """Main API endpoint for workload optimization."""

    yaml_data = request.files['yaml_file'].read()
    cluster_name = request.form['cluster_name']
    project_id = request.form['project_id']
    location = request.form['location']

    # ... (Input validation)

    try:
        yaml_info = extract_yaml_info(yaml_data)
        qos = determine_qos(yaml_info['resources']['requests'], yaml_info['resources']['limits'])

        # Check if autopilot, perform asset_inventory queries, etc.
        # ... (Implement your API logic based on the outlined steps)

        # Update YAML, call Vertex AI, etc.
        # ...

        result = {
            # ... Populate with recommended values, alerts, etc. 
        }
        return jsonify(result), 200

    except Exception as e:
        return jsonify({'error': str(e)}), 500

# ... (Rest of Flask app setup)


def read_k8_deployment_yaml(file_path):
    """
    Reads a Kubernetes YAML file and returns objects of specified kinds.

    Args:
        file_path (str): The path to the Kubernetes YAML file.
        selected_kinds (list): List of Kubernetes object kinds to include.

    Returns:
        list: List of filtered Kubernetes objects.
    """
    selected_kinds = ['Deployment', 'StatefulSet', 'Job', 'Pod', 'DaemonSet', 'CronJob']
    selected_objects = []

    try:
        with open(file_path, 'r') as file:
            docs = yaml.safe_load_all(file)
            for doc in docs:
                if doc and doc.get('kind') in selected_kinds:
                    selected_objects.append(doc)
    except FileNotFoundError:
        print(f"File not found: {file_path}")
    except yaml.YAMLError as exc:
        print(f"Error parsing YAML file: {exc}")
    return selected_objects

if __name__ == '__main__':
    project = "1034414536999"
    cluster_name ="online-boutique"
    
    k8_object = read_k8_deployment_yaml('samples/adservice.yaml')
    kind = (k8_object[0]['kind'])
    if kind == 'Deployment':
        object_details = k8_deployment_parse(k8_object[0],project,cluster_name, kind)
    
    # check if workload exists
    print(check_workload_exists(project, cluster_name, object_details))
  
    #app.run(debug=True)  # Debug mode for development
