requirements.txt
google-cloud-bigquery
google-cloud-asset
google-cloud-aiplatform
PyYAML
Flask
google-cloud-resource-manager
